backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Email
resend>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# HTTP Client
httpx>=0.27.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
